+++ CHANGES.md
@@ -93,7 +93,7 @@ It will take some time to think through and implement so im keeping the current 
 
 ## Missing features
 
-1. **No Overscan**: No buffer rows/columns for smoother scrolling
+1. ~~**No Overscan**: No buffer rows/columns for smoother scrolling~~
 2. **No Loading States**: No handling for dynamic content loading
 3. **Metrics**: For debuging puproses it would be good to see metrics of number of rendered cells / time it took to render cell etc.
 4. **Unit Tests**: Add Unit tests for most important features